requirements.txt
discord.py==2.5.2
sqlite3==3.45.3
